fixed bugs and implement new db method

+++ Source/Api/Parking.Core/Models/Data/ParkingLot.cs
@@ -8,8 +8,8 @@
 
         public int BoxCount { get; set;}
 
-        public float Longtitude { get; set;}
+        public double Longtitude { get; set;}
 
-        public float Latitude { get; set;}
+        public double Latitude { get; set;}
     }
 }
